Fix a few compiler errors.
git-svn-id: https://svn.dealii.org/trunk@28253 0785d39b-7218-0410-832d-ea1e28bc413d

+++ deal.II/source/lac/sparsity_tools.cc
@@ -163,7 +163,8 @@ namespace SparsityTools
                   break;
               // post-condition after loop: coordination, i.e. the number of
               // entries in this row is now j-rowstart[row]
-              if (j-sparsity.begin(row) <  min_coordination)
+              if (static_cast<unsigned int>(j-sparsity.begin(row)) <
+                  min_coordination)
                 {
                   min_coordination = j-sparsity.begin(row);
                   starting_point   = row;
